android/requirements.txt
lettuce
lettuce_webdriver
nose
browserstack-local
paver
selenium
psutil
Appium-Python-Client==0.52
